Refactory: Menu 관련 함수를 클래스화 및 모듈화
